Adding aboility to add half numbers in hours

+++ spec/models/invoice_spec.rb
@@ -12,7 +12,7 @@ RSpec.describe Invoice, type: :model do
   describe 'DB table' do
     it { is_expected.to have_db_column :description }
     it { is_expected.to have_db_column(:unit).of_type(:integer) }
-    it { is_expected.to have_db_column(:quantity).of_type(:integer) }
+    it { is_expected.to have_db_column(:quantity).of_type(:float) }
     it { is_expected.to have_db_column(:amount).of_type(:integer) }
     it { is_expected.to have_db_column :first_day }
     it { is_expected.to have_db_column :last_day }
